+++ ZadatakKrizicKruzic/src/main/java/hr/bojan/zadatakkrizickruzic/core/service/GameService.java
@@ -8,8 +8,8 @@ import hr.bojan.zadatakkrizickruzic.core.model.exception.IllegalActionException;
  */
 public interface GameService {
 	
-	public static final short GAME_DIFFICULTY_MAX = 4;
-	public static final short GAME_DIFFICULTY_MIN = 1;
+	public static final int GAME_DIFFICULTY_MAX = 4;
+	public static final int GAME_DIFFICULTY_MIN = 1;
 	public static final float GAME_DIFFICULTY_LIMIT_LOW = 0.3f;
 	public static final float GAME_DIFFICULTY_LIMIT_HIGH = 0.9f;
 
@@ -35,5 +35,5 @@ public interface GameService {
 	 * @return Game status after move is made
 	 * @throws IllegalActionException if illegal move is made, game has already finished, etc.
 	 */
-	public Game playGame(int gameId, short row, short column) throws IllegalActionException;
+	public Game playGame(int gameId, int row, int column) throws IllegalActionException;
 }
